dev-requirements.txt
# Flask Libraries
Flask==2.0.1
Werkzeug==2.0.1

# Others Libraries
Flask-Cors==3.0.9
python-dotenv==0.19.2


# Database Libraries
SQLAlchemy==1.4.20
PyMySQL==1.0.2



